Changing Cart subtotal to decimal

+++ src/Web/Models/Cart.cs
@@ -7,7 +7,7 @@ namespace Web.Models
     {
         public Guid CartId { get; set; }
         public Guid UserId { get; set; }
-        public double Subtotal { get; set; }
+        public decimal Subtotal { get; set; }
 
         public ICollection<CartBook> CartBooks { get; set; } = new List<CartBook>();
     }
